requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
Django==5.1.7
django-celery-results==2.5.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dotenv==0.9.9
drf-yasg==1.21.10
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
idna==3.10
inflection==0.5.1
jieba3k==0.35.1
joblib==1.4.2
kombu==5.5.2
lxml==5.3.1
lxml_html_clean==0.4.1
Markdown==3.7
newspaper3k==0.2.8
nltk==3.9.1
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
sgmllib3k==1.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.6
sqlparse==0.5.3
tinysegmenter==0.3
tldextract==5.1.3
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
